Remove disgusting ip(1) invocation and parsing
Astro uses Vite to serve it, which accepts the --host option to
serve it on the local IP address for development and testing
purposes.  No need for all this grossness to do it manually.
Simply invoking `npx astro dev --host` suffices, and furthermore it
also runs on localhost as well as the local IP address, allowing me
to test things requiring a secure context, like the Clipboard API.